Moved the rule into "identifier" package under the rewriter
